Rewrite the blackfin DMA based UART dirver to support more than one ports. The default UART port is ttyS0. User can change it in the kernel boot paramter line. More than one UART ports can be opened concurrently, this dirver can deal with each port individually.
This driver is tested on BF533 STAMP board. It also works well on the BF537
EZKIT board, but I haven't tested on it. It will be done soon.

In addition, this patch also fixes a bug which cause system hang if you enter
Ctrl+C when output a bunch of data on the console.


git-svn-id: svn://localhost/svn/uclinux-dist/trunk@962 3e2a5524-39e8-4cb9-84a9-c49da69cba3a