This patch adds an workaround for a new Blackfin Hardware anomaly:
- Only large memory system with more than 64 MB are affected.

There is an error in the hardware loop logic which can cause bad instructions to
get executed, when executing loops in the vicinity of servicing instruction ICPLB
exceptions.

The idea behind this patch is to disallow ICPLD miss exception to take place.

This patch therefore limits the maximum allowed usable instruction memory to 60 MB.
- This is what we can cover by one set of initial ICPLDs.

In order to cover as much as possible memory with ICPLDs, we moved the generic mtd
filesystem to the top of memory  this area and the uncached memory region is then not
covered by an ICPLD.


Regards, Michael


git-svn-id: svn://localhost/svn/uclinux-dist/trunk@1426 3e2a5524-39e8-4cb9-84a9-c49da69cba3a