Bug[#1568] Some gcc testcases redefine memcpy, which, causes even calls to memcpy inside uClibc to use the main program's definition.  With FD-PIC binaries, the main program isn't fully relocated until the init functions of all shared libraries have run, so if the uClibc init uses memcpy, the program may crash.
This can be fixed by having uClibc define a private hidden symbol that aliases
memcpy and calling that instead of memcpy - as a side effect that'll improve
performance since we no longer go through the PLT.

All this is done properly in current versions of uClibc.  For the moment, while
we're still using an old version, I'll use this slightly hackish fix: replace
semi-random calls to memcpy with calls to a new symbol __libc_memcpy.  That
seems to be enough to prevent uClibc startup from using the main program's
definition of memcpy.


git-svn-id: svn://localhost/svn/uclinux-dist/trunk@3393 3e2a5524-39e8-4cb9-84a9-c49da69cba3a