This fixes toolchain bug 3351, and to shut up the !@#$% pre-commit hook in the fifth attempt, I'll pretend it's actually Task[#2424].
When the shell tries to execute a file that isn't executable, the kernel
returns ENOEXEC from execve.  The shell will then try to execute it as a
shell script by spawning another DEFAULT_SHELL with the filename as first
argument.  It also sets argv[0] to the filename we're trying to execute,
which appears wrong (or at least, it can't work with a busybox shell):
consider a shell script called "test" - if we execute a shell with
argv[0] == "test", then it won't behave as a shell: the shell is busybox,
which will think it's called to perform a "test" operation.

Fixed by setting argv[0] to DEFAULT_SHELL.


git-svn-id: svn://localhost/svn/uclinux-dist/trunk@5597 3e2a5524-39e8-4cb9-84a9-c49da69cba3a