[#2424] This is actually toolchain bug 3623.
Since gcc became multilibbed, there have been multiple versions of the shared
flat libc.  When linking a shared flat executable, we must use the libc.gdb
that corresponds to the lib1.so that will eventually be installed in the root
filesystem, or mismatches can occur.  To ensure this, add $(CPUFLAGS) to the
compiler command line that is used to find libc.gdb.

This is still somewhat brittle, and probably an unintended negative consequence
of multilibbing gcc.


git-svn-id: svn://localhost/svn/uclinux-dist/trunk@5725 3e2a5524-39e8-4cb9-84a9-c49da69cba3a