+ Integrate boardenv tool
boardenv supports store/read MAC addr in /env_board.
Information fom /eeprom (obsolete) is now distributed
in /env_board and /env_user. The usage of this data
was adapted in rc script.

Individual MAC addresses are now supported. But this
approach will be replaced with the new u-boot and
shared environment variables.