+ Replaced hack for basic SPI pin reservation with proper solution
Previously, the basic SPI pin (SCK, MISO, MOSI) were configured to be
used in their function as peripheral and not GPIO pins with a hack. This
has been replaced with the proper method over the req_pin field in the
SPI configuration structure.

Because the call-order in the BF5xx SPI driver was wrong, it had to be
adjusted. This reordering was also done in the newest version of this file
in the blackfin.uclinux.org repository.